move listClusterByLabel
